Account for run_type

+++ ci/after_success.sh
@@ -6,13 +6,15 @@ set -x
 
 . ./environment
 
-HOST='user@host'
+if [ "$RUN_TYPE" = "deploy" ]; then
+    HOST='user@host'
 
-ssh "$HOST" 'rm -f shurik_new'
+    ssh "$HOST" 'rm -f shurik_new'
 
-scp ./ci/wait_deploy_run.sh "$HOST:"
-scp ./target/debug/shurik "$HOST:shurik_new"
-ssh "$HOST" 'ssh-keyscan gitlab.host > ~/.ssh/known_hosts'
+    scp ./ci/wait_deploy_run.sh "$HOST:"
+    scp ./target/debug/shurik "$HOST:shurik_new"
+    ssh "$HOST" 'ssh-keyscan gitlab.host > ~/.ssh/known_hosts'
 
-ssh "$HOST" 'sh -c "tempfile > logname"'
-ssh "$HOST" 'sh -c "nohup \"./wait_deploy_run.sh\" > $(cat logname) 2>&1 < /dev/null &"'
+    ssh "$HOST" 'sh -c "tempfile > logname"'
+    ssh "$HOST" 'sh -c "nohup \"./wait_deploy_run.sh\" > $(cat logname) 2>&1 < /dev/null &"'
+fi;
